Adding in filename checking
This begins the work of checking our input. I check if the file exists,
and I check if the filename ends with .csv. That's not enough to be
certain it is a CSV file, but it is better than nothing. We are assuming
our user is not *trying* to break the code, so it is unlikely they will
rename a PDF as .csv and force it through.

Now, an .xslx...? But, we'll proceed for now and hope that they don't do
these things.